todos comandos funcionando
