+++ functions/src/index.ts
@@ -14,19 +14,6 @@ import { registerIniciarCommand, registerStartCommand } from "./commands/start";
 import { registerAjudaCommand, registerHelpCommand } from "./commands/help";
 import { getCoordinators, getFinancesGroupId } from "./services/firebase";
 
-import { quemSouEuCommand } from "./commands/quemsoueu";
-import { pautaCommand } from "./commands/pauta";
-import { informeCommand } from "./commands/informe";
-import { demandaCommand } from "./commands/demanda";
-import { encaminhamentoCommand } from "./commands/encaminhamentos";
-import { pedidoDeInformacaoCommand } from "./commands/pedido_de_informacao";
-import { documentoCommand } from "./commands/documento";
-import { planilhaCommand } from "./commands/planilha";
-import { formularioCommand } from "./commands/formulario";
-import { apresentacaoCommand } from "./commands/apresentacao";
-import { registrarPlanilhaCommand } from "./commands/registrar_planilha";
-import { clippingCommand } from "./commands/clipping";
-
 import { eventoCommand } from "./commands/evento";
 import { pagamentoCommand } from "./commands/pagamento";
 
@@ -36,22 +23,9 @@ import { registerCancelPaymentHandler } from "./callbacks/cancelPaymentCallback"
 
 import { checkGoogleForms } from "./scheduler/checkForms";
 import { onSchedule } from "firebase-functions/scheduler";
+import { commandsList } from "./utils/commands";
 
-const validCommands = [
-  apresentacaoCommand,
-  demandaCommand,
-  documentoCommand,
-  encaminhamentoCommand,
-  formularioCommand,
-  informeCommand,
-  pautaCommand,
-  pedidoDeInformacaoCommand,
-  planilhaCommand,
-  quemSouEuCommand,
-  registrarPlanilhaCommand,
-  clippingCommand,
-];
-
+const validCommands = commandsList
 validCommands.forEach((cmd) => {
   cmd.register(bot);
 });
